requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
cssselect2==0.8.0
Django==5.1.6
django-allauth==65.4.1
django-appconf==1.1.0
django-imagekit==5.0.0
fonttools==4.56.0
idna==3.10
pilkit==3.0
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
urllib3==2.3.0
weasyprint==64.1
webencodings==0.5.1
zopfli==0.2.3.post1
